+++ lib/V8/v8-no-deadline.cpp
@@ -38,4 +38,4 @@ std::chrono::milliseconds correctTimeoutToExecutionDeadline(
 uint32_t correctTimeoutToExecutionDeadline(uint32_t timeout) { return timeout; }
 
 void TRI_InitV8Deadline(v8::Isolate* isolate) {}
-void triggerDeadlineNow() {}
+void triggerV8DeadlineNow() {}
